requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.7
psycopg2-binary>=2.9.5
pydantic>=1.10.7
pydantic-settings>=2.0.0
celery>=5.2.7
redis>=4.5.4
httpx>=0.23.3
python-dotenv>=1.0.0
alembic>=1.10.2
pytest>=7.3.1
cryptography>=40.0.1
